requirements.txt
# Core libraries
numpy==1.24.4
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1

# Natural Language Processing
spacy==3.6.0
sentence-transformers==2.2.2

# Visualization
plotly==5.15.0
umap-learn==0.5.4
streamlit==1.25.0

# HTTP requests
requests==2.31.0